fix ts-demo
