+++ package.json
@@ -72,7 +72,7 @@
   "main": "dist/joint.min.js",
   "module": "joint.mjs",
   "style": "./dist/joint.min.css",
-  "types": "./dist/index.d.ts",
+  "types": "./dist/joint.d.ts",
   "homepage": "http://jointjs.com",
   "author": {
     "name": "client IO",
