+++ test/ts/index.test.ts
@@ -1,6 +1,6 @@
 type AssertExtends<A,B> = A extends B ? true : never;
 
-import * as joint from '../../build/index';
+import * as joint from '../../build/joint';
 
 const graph = new joint.dia.Graph({ graphAttribute: true });
