+++ test/ts/mvc.listener.test.ts
@@ -1,4 +1,4 @@
-import { dia, shapes, mvc } from '../../build/index';
+import { dia, shapes, mvc } from '../../build/joint';
 
 type Args = [
     { graph: dia.Graph },
