--- types/index.d.ts
@@ -1,21 +1,0 @@
-// Definitions by:
-// Aidan Reel <http://github.com/areel>,
-// David Durman <http://github.com/DavidDurman>,
-// Ewout Van Gossum <https://github.com/DenEwout>,
-// Federico Caselli <https://github.com/CaselIT>,
-// Chris Moran <https://github.com/ChrisMoran>
-// Michael MacFadden https://github.com/mmacfadden
-
-// Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
-// typings: https://github.com/CaselIT/typings-jointjs
-
-/// <reference types="jquery" />
-/// <reference types="lodash" />
-/// <reference path="./geometry.d.ts" />
-/// <reference path="./vectorizer.d.ts" />
-/// <reference path="./mvc.d.ts" />
-/// <reference path="./joint.d.ts" />
-
-export = joint;
-export as namespace joint;
-declare namespace joint {}
